Refactor storage API in preparation for support for additional storage engines. Move GDBM-specific code into new source file.
* gzochi-server/src/app.c (initialize_data): Change names of meta, oids,
  and names database to use unqualified path names.
* gzochi-server/src/data.c (create_oid_block, dereference): int -> size_t
  for length types.
* gzochi-server/src/httpd.c (dispatch_oid, dispatch_names): int -> size_t
  for length types.
* gzochi-server/src/storage-gdbm.c: New file.
* gzochi-server/src/storage.c: Remove old GDBM storage implementation.
* gzochi-server/src/storage.h (gzochid_storage_store): Move lock metadata
  into GDBM implementation; replace GDBM_FILE with generic pointer.
  (gzochi_storage_transaction): Move cache and operation list into GDBM
  implementation; add generic pointer for engine-specific state.
  (gzochid_storage_data_lock, gzochid_storage_operation,
  gzochid_storage_operation_put): Delete GDBM-specific typedefs.
  (gzochid_storage_get, gzochid_storage_put, gzochid_storage_delete,
  gzochid_storage_first_key, gzochid_storage_next_key,
  gzochid_storage_transaction_get, gzochid_storage_transaction_put,
  gzochid_storage_transaction_delete,
  gzochid_storage_transaction_first_key,
  gzochid_storage_transaction_next_key): int -> size_t for length types.

+++ gzochi-server/src/data.c
@@ -198,7 +198,7 @@ static gzochid_data_oid_block *create_oid_block
 {
   char next_oid_key[] = { NEXT_OID_KEY };
   char *next_oid_value = NULL;
-  int next_oid_value_len = 0;
+  size_t next_oid_value_len = 0;
 
   gzochid_data_oid_block *block = calloc (1, sizeof (gzochid_data_oid_block));
 
@@ -388,7 +388,7 @@ void dereference
 (gzochid_data_transaction_context *context, 
  gzochid_data_managed_reference *reference)
 {
-  int data_len = 0;
+  size_t data_len = 0;
   char *oid_str = NULL; 
   char *data = NULL; 
   GString *in = NULL;
